+++ package.json
@@ -50,9 +50,9 @@
     "@fastify/cors": "^8.2.0",
     "@fastify/swagger": "^7.6.1",
     "@fastify/type-provider-typebox": "^3.2.0",
-    "@hirosystems/chainhook-client": "^1.0.2",
+    "@hirosystems/api-toolkit": "^1.4.0",
+    "@hirosystems/chainhook-client": "file:../chainhook/components/client/typescript",
     "@sinclair/typebox": "^0.28.17",
-    "@hirosystems/api-toolkit": "^1.0.0",
     "@stacks/transactions": "^6.1.0",
     "@types/node": "^18.11.9",
     "env-schema": "^5.1.0",
